cleanup gen command
