+++ spraycharles/lib/utils/__init__.py
@@ -1,4 +1,3 @@
-from spraycharles.lib.utils.make_list import make_list
 from spraycharles.lib.utils.notify import discord, teams, slack, HookSvc
 from spraycharles.lib.utils.ntlm_challenger import main as ntlm_challenger
 from spraycharles.lib.utils.smbstatus import SMBStatus
